requirements.txt
django==1.11.6
djangorestframework
django-crispy-forms
pytesseract

